memory/requirements.txt
fastapi
uvicorn
mem0ai
ollama
pydantic
python-dotenv
pytest
httpx
qdrant-client